Vy push code quan ly user va quan ly hoa don

+++ database/migrations/2026_01_05_150939_create_sanpham_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->string('tensp');
             $table->string('hinhanh')->nullable();
             $table->integer('soluong');
-            $table->decimal('giaban');
+            $table->double('giaban');
             $table->string('slug');
             $table->string('mota')->nullable();
             $table->foreignId('bienthe_id')->constrained('bienthe')->cascadeOnDelete();
